Add files via upload
Problem Definition:
The task here is to predict whether a bank currency note is authentic or not based on four attributes i.e. variance of the image wavelet transformed image, skewness, entropy, and curtosis of the image.
Solution:
We will use a random forest and support vector machine to solve this problem in terms of performance and conclude which algorithm gives more accuracy.